requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
tzdata==2025.1
djangorestframework>=3.13.1
requests>=2.28.1
isort
flake8
